add makefile

+++ 2023/day02/cpp/main.cpp
@@ -44,7 +44,7 @@ int main() {
     int green_max = 0;
     int blue_max = 0;
 
-    for (int i = 0; i < line.size(); ++i) {
+    for (size_t i = 0; i < line.size(); ++i) {
       char c = line[i];
 
       if (c == ',' || c == ';' || i == line.size() - 1) {
